# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlphaPilot API",
    "description": "AI Autonomous Trading System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/decisions": {
      "get": {
        "tags": [
          "decisions"
        ],
        "summary": "List Decisions",
        "description": "返回最近 N 条 AI 决策记录 (要求登录)。",
        "operationId": "list_decisions",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "ap_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ap Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_list_DecisionRead__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/decisions/{decision_id}": {
      "get": {
        "tags": [
          "decisions"
        ],
        "summary": "Get Decision Detail",
        "description": "决策详情: features 快照 + review + 守卫记录 + 关联订单 (要求登录)。",
        "operationId": "get_decision_detail",
        "parameters": [
          {
            "name": "decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Decision Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "ap_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ap Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_DecisionDetailOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/risk/state": {
      "get": {
        "tags": [
          "risk"
        ],
        "summary": "Get Risk State",
        "description": "风控状态快照: OK|WARN|HALTED + 日亏/仓位占比/regime (要求登录)。",
        "operationId": "get_risk_state",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "ap_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ap Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_RiskStateOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionDetailOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "decided_at": {
            "type": "string",
            "title": "Decided At"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "entry_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Type"
          },
          "entry_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Price"
          },
          "stop_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss"
          },
          "take_profit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit"
          },
          "position_size_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Size Pct"
          },
          "strategy_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Mode"
          },
          "reasoning": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "risk_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Note"
          },
          "is_fallback": {
            "type": "boolean",
            "title": "Is Fallback"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "llm_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Provider"
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "tokens_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens Used"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "features": {
            "$ref": "#/components/schemas/DecisionFeaturesOut"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/DecisionReviewRead"
            },
            "type": "array",
            "title": "Reviews"
          },
          "guard_events": {
            "items": {
              "$ref": "#/components/schemas/DecisionGuardEventRead"
            },
            "type": "array",
            "title": "Guard Events"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/DecisionOrderRead"
            },
            "type": "array",
            "title": "Orders"
          },
          "position_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Position Ids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "timeframe",
          "decided_at",
          "action",
          "is_fallback",
          "features",
          "reviews",
          "guard_events",
          "orders",
          "position_ids"
        ],
        "title": "DecisionDetailOut"
      },
      "DecisionFeaturesOut": {
        "properties": {
          "factor_snapshot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Factor Snapshot Id"
          },
          "factors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Factors"
          },
          "factor_def_versions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Factor Def Versions"
          },
          "prompt_input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Input"
          }
        },
        "type": "object",
        "title": "DecisionFeaturesOut"
      },
      "DecisionGuardEventRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "triggered_at": {
            "type": "string",
            "title": "Triggered At"
          },
          "resolved": {
            "type": "boolean",
            "title": "Resolved"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_type",
          "description",
          "triggered_at",
          "resolved"
        ],
        "title": "DecisionGuardEventRead"
      },
      "DecisionOrderRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "side": {
            "type": "string",
            "title": "Side"
          },
          "order_type": {
            "type": "string",
            "title": "Order Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "avg_fill_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Fill Price"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "filled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filled At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "side",
          "order_type",
          "status",
          "trace_id"
        ],
        "title": "DecisionOrderRead"
      },
      "DecisionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "strategy_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Mode"
          },
          "reasoning": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "risk_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Note"
          },
          "is_fallback": {
            "type": "boolean",
            "title": "Is Fallback"
          },
          "decided_at": {
            "type": "string",
            "title": "Decided At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "timeframe",
          "action",
          "is_fallback",
          "decided_at"
        ],
        "title": "DecisionRead"
      },
      "DecisionReviewRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "reviewer_type": {
            "type": "string",
            "title": "Reviewer Type"
          },
          "result": {
            "type": "string",
            "title": "Result"
          },
          "adjustments": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjustments"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reviewer_type",
          "result"
        ],
        "title": "DecisionReviewRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Response_DecisionDetailOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "成功"
          },
          "detailMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailmessage"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionDetailOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "Response[DecisionDetailOut]"
      },
...
      "Response_RiskStateOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "成功"
          },
          "detailMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailmessage"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskStateOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "Response[RiskStateOut]"
      },
...
      "Response_list_DecisionRead__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "成功"
          },
          "detailMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailmessage"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DecisionRead"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "Response[list[DecisionRead]]"
      },
...
      "RiskStateOut": {
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "OK",
              "WARN",
              "HALTED"
            ],
            "title": "State"
          },
          "day_loss_pct": {
            "type": "number",
            "title": "Day Loss Pct"
          },
          "positions_pct": {
            "type": "number",
            "title": "Positions Pct"
          },
          "regime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regime"
          }
        },
        "type": "object",
        "required": [
          "state",
          "day_loss_pct",
          "positions_pct"
        ],
        "title": "RiskStateOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}